Adicionado mais exercicios

+++ exercicios/conversorCm.py
@@ -1,3 +1,3 @@
-num=int (input("Digite um numero em centimetros: "))
+num=float (input("Digite um numero em centimetros: "))
 mm=num*10
 print("O numero {} em milimetros é: {}".format(num,mm))
