+++ exercicios/conversorReal.py
@@ -1,3 +1,3 @@
-num=int (input("Digite um numero em reais: "))
+num=float (input("Digite um numero em reais: "))
 dolar=num/5
 print("o valor em dolares foi : {}".format(dolar))
