+++ exercicios/media.py
@@ -1,6 +1,5 @@
 nome=input("Informe o nome do aluno:  ")
-nota1=int(input("Informe a nota 1: "))
-nota2=int(input("Informe a nota 2: "))
+nota1=float(input("Informe a nota 1: "))
+nota2=float(input("Informe a nota 2: "))
 media=(nota1+nota2)/2
-
 print("A media das notas {} e {} é {}".format(nota1, nota2, media))
